Change `figure` schema to work better with GatsbyImage
GraphQL images in arrays don't work nicely with custom Sanity objects when they're nested not as an `object` but a custom type `figure`. By changing the schema to have a base type `image` nested in a `figure` type `object`, everything works properly. This sets up all the pre-existing pages to use the correct schema (and I'm leaving them in currently for reference and as a resource).