Don't allow drafts in references!
Good lord, that caused a lot of confusion on the GraphQL end. Don't allow users to add drafted reference documents, because it causes a lot of weirdness when trying to save them after the fact. Not to mention that none of those drafts show up in the front-end, which then causes things to break.