Remove `fit=crop` from image URL building
We never wanted this, it turns out. It will crop the image in unexpected ways when there is a crop coming from the admin. I'm still not sure why it was cropping the homepage image so aggressively, but removing this lets the image fill in only using the admin-set crop, and then pass the whole image through. I tested this on product page crops, and it doesn't cause any issues there.